+++ R/a_summarize_aval_chg_diff.R
@@ -477,8 +477,8 @@ a_summarize_aval_chg_diff_j <- function(
   ctrl_grp <- NULL
   if (comp_btw_group) {
     ref <- get_ref_info(ref_path, .spl_context)
-    split_var_refspec <- ref$split_var
-    checkmate::assert_true(identical(trt_var, split_var_refspec))
+    split_name_refspec <- ref$split_name
+    checkmate::assert_true(identical(trt_var, split_name_refspec))
     ctrl_grp <- ref$ref_level
 
     ### check that ctrl_grp is a level of the treatment variable, in case riskdiff is requested