+++ NEWS.md
@@ -40,7 +40,7 @@
 - Update new exported calls from rtables.officer
 - update documentation to `roxygen2` 8.0.0 
 - Add extra statistics to `a_eair100_j` and introduce scaling factor `num_p_year` (default = 100) (#361)
-- Unified `get_ref_info()` which now also returns `split_var`, `ref_level`, and `cur_col_val` (#295)
+- Unified `get_ref_info()` which now also returns `split_name`, `ref_level`, and `cur_col_val` (#295)
 - `h_get_trtvar_refpath()` is marked as superseded
 - `a_summarize_aval_chg_diff_j()` now uses `get_ref_info()`